// store/auth/types.ts
export enum MutationsTypes {
  SET_USER = 'setUser',
}

export enum ActionsTypes {
  SIGN_IN = 'signIn',
  SIGN_IN_WITH_GOOGLE = 'signInWithGoogle',
  SIGN_IN_WITH_FACEBOOK = 'signInWithFacebook',
  SIGN_UP = 'signUp',
  SET_SESSION = 'setSession',
}

export enum SignInProviders {
  EMAIL = 'email',
  FACEBOOK = 'facebook',
  GOOGLE = 'google',
}